requirements.txt
tensorflow>=2.11
numpy
scikit-learn
pyyaml
matplotlib
pandas